remix_full_without_measurements/openai-api-garment-attribute-predictor/requirements.txt
# Web framework and server
Flask==2.3.3
gunicorn==21.2.0

# Machine Learning and AI
torch==2.0.1
transformers==4.35.2
openai==1.51.0

# Image processing
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3

# AWS services
boto3==1.29.7

# HTTP requests and utilities
requests==2.31.0
httpx==0.27.0

# Environment management
python-dotenv==1.0.0

# Additional utilities that might be needed
urllib3==1.26.18
